+++ identity/src/main/java/com/stripe/android/identity/StripeIdentityVerificationSheet.kt
@@ -1,0 +1,11 @@
+package com.stripe.android.identity
+
+class StripeIdentityVerificationSheet : IdentityVerificationSheet {
+    override fun present(
+        verificationSessionId: String,
+        ephemeralKeySecret: String,
+        onFinished: (verificationResult: IdentityVerificationSheet.VerificationResult) -> Unit
+    ) {
+        onFinished(IdentityVerificationSheet.VerificationResult.Completed)
+    }
+}