+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/paymentdatacollection/CardDataCollectionFragment.kt
@@ -358,7 +358,7 @@ internal class CardDataCollectionFragment : Fragment() {
         requireArguments().getParcelable<FormFragmentArguments>(
             ComposeFormDataCollectionFragment.EXTRA_CONFIG
         )?.let { args ->
-            saveCardCheckbox.isChecked = true
+            saveCardCheckbox.isChecked = false
             saveCardCheckbox.isVisible = args.showCheckbox
         }
         sheetViewModel.newCard?.customerRequestedSave?.also {