+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/paymentdatacollection/CardDataCollectionFragmentTest.kt
@@ -187,7 +187,7 @@ internal class CardDataCollectionFragmentTest : PaymentSheetViewModelTestInjecti
             assertThat(viewBinding.billingAddress.countryView.text.toString())
                 .isEqualTo("Germany")
             assertThat(viewBinding.saveCardCheckbox.isVisible).isTrue()
-            assertThat(viewBinding.saveCardCheckbox.isChecked).isTrue()
+            assertThat(viewBinding.saveCardCheckbox.isChecked).isFalse()
         }
     }